ml-flask-backend-main/ml-flask-backend-main/requirements.txt
Flask
flask-cors
tensorflow
numpy
opencv-python
Pillow
# # 1️⃣ Install dependencies
# pip install -r requirements.txt

# 2️⃣ Train model (creates model.h5)
# python train_model.py

# 3️⃣ Run Flask server
# python run.py
